% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tempconverter.R
\name{C_to_F}
\alias{C_to_F}
\title{Santhosh's Temp Converter}
\usage{
C_to_F(C_temp)
}
\arguments{
\item{C_temp}{Temp value in Celsius}
}
\value{
Temp Converted to Fahrenheit
}
\description{
Temperature conversion between Celsius and Fahrenheit,
    can convert both ways.
}
\examples{
C_to_F(95)

}
