% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tempconverter.R
\name{F_to_C}
\alias{F_to_C}
\title{Santhosh's Temp Converter}
\usage{
F_to_C(F_temp)
}
\arguments{
\item{F_temp}{Temp value in Fahrenheit}
}
\value{
Temp Converted to Celsius
}
\description{
Temperature conversion between Celsius and Fahrenheit,
    can convert both ways.
}
\examples{
F_to_C(95)

}
